requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
# torch and torchaudio are installed in Dockerfile with CUDA wheels
transformers==4.46.1
git+https://github.com/huggingface/parler-tts.git
soundfile==0.12.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.2
librosa==0.10.2.post1
numpy==1.26.4
onnxruntime
onnxruntime-gpu
celery==5.3.6
redis==5.0.1
torchcodec

# RAG Pipeline Dependencies
qdrant-client==1.7.0
sentence-transformers==2.2.2
pypdf==3.17.4
python-docx==1.1.0
pillow==10.1.0
pytesseract==0.3.10
easyocr==1.7.0
langchain==0.1.0
langchain-community==0.0.10
ragas==0.1.0
datasets==2.16.1